+++ ainigma-backend/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "ainigma-backend"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+ainigma = { path = "../ainigma" }
+
+# and any other dependencies you need:
+hyper = "1.6.0"
+axum = "0.8.4"
+tokio = { version = "1", features = ["full"] }
+uuid = { version = "1.17.0", features = ["v7", "serde"] }
+serde = { version = "1.0.211", features = ["derive"] }
+serde_json = "1"
+tower = "0.5.2"
+sqlx = { version = "0.8.6", features = ["postgres", "runtime-tokio-rustls"] }
+
+[dev-dependencies]
+reqwest = { version = "0.12.19", features = ["json"] }
+tokio = { version = "1", features = ["full"] }
