+++ ainigma/src/errors/base.rs
