+++ ainigma/src/flag_generator.rs
