+++ ainigma/tests/tasks/reverse1/custom_entry.sh
