+++ ainigma/tests/tasks/reverse1/entrypoint.sh
