+++ ainigma/tests/tasks/simple_shell/build.sh
